requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
lxml==6.0.2
markupsafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
ruff==0.12.4
soupsieve==2.8
typing-extensions==4.15.0
urllib3==2.5.0
validators==0.35.0
werkzeug==3.1.3
